Changed to write test results to a file
There is probably an easier way to do this, but I got it working and the code appends three different results files, one for each algorithm. This was achieved by changing printMetrics to a public int[] that returns an array of the results. The code still works the same otherwise.